requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-login
flask-bcrypt
flask-mail
Pillow
